+++ hs-bindgen-runtime/test/Test/HsBindgen/Runtime/Support/ByteArray.hs
@@ -20,8 +20,13 @@ import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.QuickCheck (testProperty)
 import Text.Printf (printf)
 
-import HsBindgen.Runtime.Support.ByteArray (getUnionPayload, setUnionPayload)
+import HsBindgen.Runtime.Support.Bitfield (Bitfield (narrow))
+import HsBindgen.Runtime.Support.ByteArray (getUnionPayload,
+                                            getUnionPayloadBits,
+                                            setUnionPayload,
+                                            setUnionPayloadBits)
 
+import Test.Util.QC (Arbitrary4 (liftShrink4))
 import Test.Util.Show (showRangesOf)
 
 {-------------------------------------------------------------------------------
@@ -43,6 +48,20 @@ tests = testGroup "Test.HsBindgen.Runtime.Support.ByteArray" [
     , testProperty "prop_getSet @Word32" $ prop_getSet (Proxy @Word32)
     , testProperty "prop_getSet @Word64" $ prop_getSet (Proxy @Word64)
     , testProperty "prop_getSet @Int"    $ prop_getSet (Proxy @Int)
+
+      -- prop_setGetBits
+    , testProperty "prop_setGetBits @Word8"  $ prop_setGetBits (Proxy @Word8)
+    , testProperty "prop_setGetBits @Word16" $ prop_setGetBits (Proxy @Word16)
+    , testProperty "prop_setGetBits @Word32" $ prop_setGetBits (Proxy @Word32)
+    , testProperty "prop_setGetBits @Word64" $ prop_setGetBits (Proxy @Word64)
+    , testProperty "prop_setGetBits @Int"    $ prop_setGetBits (Proxy @Int)
+
+      -- prop_getSetBits
+    , testProperty "prop_getSetBits @Word8"  $ prop_getSetBits (Proxy @Word8)
+    , testProperty "prop_getSetBits @Word16" $ prop_getSetBits (Proxy @Word16)
+    , testProperty "prop_getSetBits @Word32" $ prop_getSetBits (Proxy @Word32)
+    , testProperty "prop_getSetBits @Word64" $ prop_getSetBits (Proxy @Word64)
+    , testProperty "prop_getSetBits @Int"    $ prop_getSetBits (Proxy @Int)
     ]
 
 {-------------------------------------------------------------------------------
@@ -142,6 +161,122 @@ instance (Storable a, Arbitrary a, Show a) => Arbitrary (SetGetParams a) where
           ]
 
 {-------------------------------------------------------------------------------
+  Roundtrip: set bits and get bits
+-------------------------------------------------------------------------------}
+
+-- | Roundtrip: get bits after set bits should return the set value
+prop_setGetBits ::
+     forall a. Bitfield a
+  => Proxy a -> SetGetBitsParams a -> Property
+prop_setGetBits _ params =
+    tabulate "# of unused leading bits" [showRangesOf 64 o] $
+    tabulate "# of unused trailing bits" [showRangesOf 64 (bsSz * 8 - (o + w))] $
+    narrow v w === getUnionPayloadBits o w (setUnionPayloadBits o w v bs)
+  where
+    o = params.bitOffset
+    w = params.bitWidth
+    v = params.value
+    bs = params.bytes
+    bsSz = P.sizeofByteArray bs
+
+-- | Roundtrip: set after get should return the byte array to its original state
+--
+-- See als 'prop_getSet'.
+prop_getSetBits ::
+     forall a. Bitfield a
+  => Proxy a -> SetGetBitsParams a -> Property
+prop_getSetBits _ params =
+    tabulate "# of unused leading bits" [showRangesOf 64 o] $
+    tabulate "# of unused trailing bits" [showRangesOf 64 (bsSz * 8 - (o + w))] $
+    bs === setUnionPayloadBits @a o w (getUnionPayloadBits o w bs) bs
+  where
+    o = params.bitOffset
+    w = params.bitWidth
+    bs = params.bytes
+    bsSz = P.sizeofByteArray bs
+
+
+-- | Parameters for 'prop_setGetBits' and 'prop_getSetBits'
+--
+-- INVARIANT: see the 'CheckInvariant' instance
+data SetGetBitsParams a = SetGetBitsParams {
+    value :: a
+  , bitOffset :: Int
+  , bitWidth :: Int
+  , bytes :: ByteArray
+  }
+  deriving stock (Show, Eq)
+
+instance (Show a, Storable a) => CheckInvariant (SetGetBitsParams a) where
+  checkInvariant params = first (printf "For params (%s), " (show params) ++) go
+    where
+      go
+        | o < 0
+        = Left $ printf "A: %d < 0" o
+        | w < 1 || w > 64
+        = Left $ printf "B: %d < 1 || %d > 64" w w
+        | let vSz = sizeOf v
+        , w > vSz * 8
+        = Left $ printf "C: %d + %d > %d * 8" o w vSz
+        | let bsSz = P.sizeofByteArray bs
+        , o + w > bsSz * 8
+        = Left $ printf "D: %d + %d > %d * 8" o w bsSz
+        | otherwise
+        = Right params
+
+      o = params.bitOffset
+      w = params.bitWidth
+      v = params.value
+      bs = params.bytes
+
+mkSetGetBitsParams ::
+     forall a. (Storable a, Show a)
+  => a -> Int -> Int -> ByteArray -> Either String (SetGetBitsParams a)
+mkSetGetBitsParams value bitOffset bitWidth bytes =
+    checkInvariant $
+      SetGetBitsParams {
+          value = value, bitOffset = bitOffset, bitWidth = bitWidth, bytes = bytes
+        }
+
+mkSetGetBitsParams' ::
+     forall a. (HasCallStack, Storable a, Show a)
+  => a -> Int -> Int -> ByteArray ->  SetGetBitsParams a
+mkSetGetBitsParams' value bitOffset bitWidth bytes =
+    checkInvariant'
+      SetGetBitsParams {
+          value = value, bitOffset = bitOffset, bitWidth = bitWidth, bytes = bytes
+        }
+
+instance (Storable a, Show a, Arbitrary a) => Arbitrary (SetGetBitsParams a) where
+  arbitrary = do
+      bitWidth <- chooseInt (1, valueSz * 8 - 1)
+      let i = ceilDiv8 bitWidth
+      bytes <- genBytes i
+      bitOffset <- chooseInt (0, P.sizeofByteArray bytes * 8 - bitWidth)
+      value <- arbitrary
+      pure $ mkSetGetBitsParams' value bitOffset bitWidth bytes
+   where
+      valueSz = sizeOf (undefined :: a)
+
+      genBytes i = sized $ \n -> do
+        k <- chooseInt (i, max i n)
+        let genByte = getLarge <$> arbitrary
+        byteArrayOf k genByte
+
+  shrink params = snd $ partitionEithers [
+        mkSetGetBitsParams value' bitOffset' bitWidth' bytes'
+      | (value', bitOffset', bitWidth', bytes')
+          <- liftShrink4 shrink shrink shrink shrinkBytes
+              (params.value, params.bitOffset, params.bitWidth, params.bytes)
+      ]
+    where
+      shrinkBytes bytes = [
+            bytes'
+          | let shrinkByte = fmap getLarge . shrink . Large
+          , bytes' <- shrinkByteArray shrinkByte bytes
+          ]
+
+{-------------------------------------------------------------------------------
   Invariants
 -------------------------------------------------------------------------------}
 
@@ -172,3 +307,11 @@ shrinkByteArray shrinkByte bytes = [
       fromList bytes'
     | bytes' <- shrinkList shrinkByte (toList bytes)
     ]
+
+{-------------------------------------------------------------------------------
+  Numeric
+-------------------------------------------------------------------------------}
+
+-- | Divide by 8 and round up
+ceilDiv8 :: Int -> Int
+ceilDiv8 x = (x + 7) `div` 8